+++ source/source/index.ts
@@ -1,1 +1,2 @@
+export { SourceFile } from "./SourceFile.js";
 export { SourceService } from "./SourceService.js";
